Use common Trackable base interface for Resources and Repositories
This makes it possible to handle resources and repositories with the
same code.

The really important change from a user perspective is that it is
possible to use modules as command line scripts (using either the
module's file name or module name, and the module will be cached
so that it isn't evaluated a second time when it is imported by
another module.